move constants to dedicated constants file
