+++ packages/back-end/src/constants/stats.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_REGRESSION_ADJUSTMENT_DAYS = 14;
+export const DEFAULT_SEQUENTIAL_TESTING_TUNING_PARAMETER = 5000;
